added csv parser

+++ src/internal/db/structs.go
@@ -1,8 +1,8 @@
 package db
 
 type RenewablesAPIData struct {
-	Name       string `json:"name"`
-	ISO        string `json:"isoCode"`
-	Year       string `json:"year,omitempty"`
-	Percentage int    `json:"percentage"`
+	Name       string  `json:"name"`
+	ISO        string  `json:"isoCode"`
+	Year       string  `json:"year,omitempty"`
+	Percentage float64 `json:"percentage"`
 }
